requirements.txt
tensorflow>=2.13.0
numpy<2
matplotlib>=3.5.0
scikit-learn>=1.0.0
pandas>=1.3.0
Pillow>=9.0.0 